feat: rewrite tests to typescript
