+++ test/Radar.spec.ts
@@ -1,20 +1,19 @@
 import { mount } from '@vue/test-utils'
-import BubbleChart from './examples/BubbleChart.vue'
+import RadarChart from './examples/RadarChart'
 
-describe('BubbleChart', () => {
+describe('RadarChart', () => {
   const Component = {
-    template:
-      '<div><BubbleChart :chartId="chartId" :plugins="plugins" /></div>',
-    components: { BubbleChart },
+    template: '<div><RadarChart :chartId="chartId" :plugins="plugins" /></div>',
+    components: { RadarChart },
     props: ['chartId', 'plugins']
   }
 
   it('should render a canvas', () => {
     const wrapper = mount(Component)
 
-    const bubbleChartEl = wrapper.find('#bubble-chart')
-    expect(bubbleChartEl.element.id).not.toBe('undefined')
-    expect(bubbleChartEl.exists()).toBe(true)
+    const radarChartEl = wrapper.find('#radar-chart')
+    expect(radarChartEl.element.id).not.toBe('undefined')
+    expect(radarChartEl.exists()).toBe(true)
 
     const canvasEl = wrapper.find('canvas')
     expect(canvasEl.exists()).toBe(true)
@@ -22,12 +21,12 @@ describe('BubbleChart', () => {
 
   it('should change id based on prop', () => {
     const wrapper = mount(Component, {
-      props: { chartId: 'bubblechartprop' }
+      props: { chartId: 'rodarchartprop' }
     })
 
-    const bubbleChartEl = wrapper.find('#bubblechartprop')
-    expect(bubbleChartEl.element.id).not.toBe('undefined')
-    expect(bubbleChartEl.exists()).toBe(true)
+    const radarChartEl = wrapper.find('#rodarchartprop')
+    expect(radarChartEl.element.id).not.toBe('undefined')
+    expect(radarChartEl.exists()).toBe(true)
   })
 
   it('should add inline plugins based on prop', () => {
