+++ test/examples/BarChart.ts
@@ -1,27 +1,29 @@
-<script>
-import { defineComponent, h } from 'vue'
+import { defineComponent, h, PropType } from 'vue'
 
-import { Pie } from '../../src/index'
+import { Bar } from '../../src'
+
 import {
   Chart as ChartJS,
   Title,
   Tooltip,
   Legend,
-  ArcElement,
-  CategoryScale
+  BarElement,
+  CategoryScale,
+  LinearScale,
+  PluginOptionsByType
 } from 'chart.js'
 
-ChartJS.register(Title, Tooltip, Legend, ArcElement, CategoryScale)
+ChartJS.register(Title, Tooltip, Legend, BarElement, CategoryScale, LinearScale)
 
 export default defineComponent({
-  name: 'PieChart',
+  name: 'BarChart',
   components: {
-    Pie
+    Bar
   },
   props: {
     chartId: {
       type: String,
-      default: 'pie-chart'
+      default: 'bar-chart'
     },
     width: {
       type: Number,
@@ -36,21 +38,35 @@ export default defineComponent({
       type: String
     },
     styles: {
-      type: Object,
+      type: Object as PropType<Partial<CSSStyleDeclaration>>,
       default: () => {}
     },
     plugins: {
-      type: Object,
+      type: Object as PropType<PluginOptionsByType<'bar'>>,
       default: () => {}
     }
   },
   setup(props) {
     const chartData = {
-      labels: ['VueJs', 'EmberJs', 'ReactJs', 'AngularJs'],
+      labels: [
+        'January',
+        'February',
+        'March',
+        'April',
+        'May',
+        'June',
+        'July',
+        'August',
+        'September',
+        'October',
+        'November',
+        'December'
+      ],
       datasets: [
         {
-          backgroundColor: ['#41B883', '#E46651', '#00D8FF', '#DD1B16'],
-          data: [40, 20, 80, 10]
+          label: 'Data One',
+          backgroundColor: '#f87979',
+          data: [40, 20, 12, 39, 10, 40, 39, 80, 40, 20, 12, 11]
         }
       ]
     }
@@ -61,7 +77,7 @@ export default defineComponent({
     }
 
     return () =>
-      h(Pie, {
+      h(Bar, {
         chartData,
         chartOptions,
         chartId: props.chartId,
@@ -73,4 +89,3 @@ export default defineComponent({
       })
   }
 })
-</script>
